Project/Phase2/requirements.txt
anyio==4.3.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asttokens==2.4.1
attrs==23.2.0
Babel==2.14.0
beautifulsoup4==4.12.3
bleach==6.1.0
cachetools==5.5.2
certifi==2024.2.2
cffi==1.17.1
charset-normalizer==2.0.12
click==8.1.7
colorama==0.4.6
comm==0.2.2
contourpy==1.2.1
cycler==0.12.1
debugpy==1.8.1
decorator==5.1.1
defusedxml==0.7.1
entrypoints==0.4
et_xmlfile==2.0.0
exceptiongroup==1.2.0
executing==2.0.1
fastjsonschema==2.19.1
fasttext-wheel==0.9.2
filelock==3.18.0
FLAML==1.0.14
flashtext==2.7
fonttools==4.51.0
fqdn==1.5.1
fsspec==2025.3.2
future==0.18.2
gensim==4.3.2
google-analytics-data==0.18.17
google-api-core==2.24.2
google-auth==2.38.0
googleapis-common-protos==1.69.1
greenlet==3.1.1
grpcio==1.71.0
grpcio-status==1.71.0
h11==0.14.0
h2o==3.36.1.4
hazm==0.10.0
huggingface-hub==0.30.2
idna==3.7
imbalanced-learn==0.12.2
importlib_metadata==7.1.0
ipykernel==6.29.4
ipython==8.18.1
ipython-genutils==0.2.0
ipywidgets==8.1.2
isoduration==20.11.0
jalali_core==1.0.0
jdatetime==5.2.0
jedi==0.19.1
Jinja2==3.1.3
joblib==1.2.0
json5==0.9.25
jsonpointer==2.4
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
jupyter==1.0.0
jupyter-console==6.6.3
jupyter-events==0.10.0
jupyter_client==7.4.9
jupyter_core==5.7.2
jupyter_server==2.14.0
jupyter_server_terminals==0.5.3
jupyterlab==3.5.1
jupyterlab_pygments==0.3.0
jupyterlab_server==2.26.0
jupyterlab_widgets==3.0.10
Khayyam==3.0.17
kiwisolver==1.4.5
Levenshtein==0.25.1
lightgbm==4.3.0
MarkupSafe==2.1.5
matplotlib==3.6.2
matplotlib-inline==0.1.6
mistune==3.0.2
mpmath==1.3.0
mysql-connector-python==9.2.0
nbclassic==1.0.0
nbclient==0.10.0
nbconvert==7.16.3
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.2.1
nltk==3.8.1
notebook==6.5.6
notebook_shim==0.2.4
numpy==1.24.3
openpyxl==3.1.5
outcome==1.3.0.post0
overrides==7.7.0
packaging==24.0
pandas==2.2.3
pandocfilters==1.5.1
parso==0.8.4
persiantools==5.2.0
pickle-mixin==1.0.2
pillow==10.3.0
platformdirs==4.2.0
plotly==5.22.0
prometheus_client==0.20.0
prompt-toolkit==3.0.43
proto-plus==1.26.1
protobuf==5.29.3
psutil==5.9.8
psycopg2==2.9.10
pure-eval==0.2.2
pyasn1==0.6.1
pyasn1_modules==0.4.1
pybind11==2.12.0
pycparser==2.22
Pygments==2.17.2
PyMySQL==1.1.1
pyodbc==5.2.0
pyparsing==3.1.2
PySocks==1.7.1
python-crfsuite==0.9.10
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-json-logger==2.0.7
pytz==2024.1
PyYAML==6.0.1
pyzmq==24.0.1
qtconsole==5.5.1
QtPy==2.4.1
rapidfuzz==3.9.4
raylib==5.5.0.2
rbo==0.1.2
referencing==0.34.0
regex==2024.5.15
requests==2.32.3
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rpds-py==0.18.0
rsa==4.9
safetensors==0.5.3
scikit-learn==1.2.2
scipy==1.8.0
seaborn==0.12.1
selenium==4.29.0
Send2Trash==1.8.3
sentence-transformers==4.1.0
six==1.16.0
smart-open==7.0.4
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.5
SQLAlchemy==1.4.54
stack-data==0.6.3
sympy==1.14.0
tabulate==0.8.10
tenacity==8.2.3
terminado==0.18.1
threadpoolctl==3.4.0
tinycss2==1.2.1
tokenizers==0.21.1
tomli==2.0.1
torch==2.7.0
tornado==6.4
tqdm==4.66.4
traitlets==5.14.2
transformers==4.51.3
trio==0.29.0
trio-websocket==0.12.2
types-python-dateutil==2.9.0.20240316
typing_extensions==4.11.0
tzdata==2025.2
uri-template==1.3.0
urllib3==1.26.18
wcwidth==0.2.13
webcolors==1.13
webencodings==0.5.1
websocket-client==1.8.0
widgetsnbextension==4.0.10
wrapt==1.16.0
wsproto==1.2.0
xgboost==1.7.2
zipp==3.18.1